spotbit-plugin/requirements.txt
cryptoadvance.specter==1.8.1
ccxt>=1.74.49